requirements.txt
anthropic>=0.18.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apify-client>=1.6.0
mcporter>=0.1.0
gunicorn>=21.2.0
